[BE/feat] 필요 칼로리들을 계산하여 반환하는 유스케이스 구현

+++ BE/exceed/src/main/java/com/gaebaljip/exceed/dto/CreateMemberRequest.java
@@ -4,9 +4,9 @@ import com.gaebaljip.exceed.member.domain.Activity;
 import lombok.Builder;
 
 public record CreateMemberRequest(
-        Integer height,
+        Double height,
         Boolean gender,
-        Integer weight,
+        Double weight,
         Integer age,
         Activity activity,
         String etc
